pyproject.toml
[project]
name = "ax-prover"
dynamic = ["version"]
description = "LangGraph agent for Lean4 theorem formalization and proving"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [
    { name = "Borja Requena", email = "borja@axiomatic-ai.com" },
    { name = "Austin Letson", email = "austin@axiomatic-ai.com" },
    { name = "Krystian Nowakowski", email = "krystian@axiomatic-ai.com" },
    { name = "Izan Beltran Ferreiro", email = "izan@axiomatic-ai.com" },
    { name = "Leopoldo Sarra", email = "leopoldo@axiomatic-ai.com" },
]
keywords = ["lean4", "theorem-proving", "formalization", "langgraph", "mathlib"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "aiohttp>=3.9.0",
    "google-auth>=2.23.0",
    "langgraph>=1.0.3",
    "langchain>=1.0.8",
    "langchain-anthropic>=1.3.0",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "lean-interact>=0.1.0",
    "jsonargparse[omegaconf]>=4.35.0",
    "omegaconf>=2.3.0",
    "platformdirs>=4.0",
    "pydantic>=2.12.4",
    "pylatexenc>=2.10",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "tavily-python>=0.7.13",
    "wikipedia>=1.4.0",
]

[project.urls]
Homepage = "https://github.com/Axiomatic-AI/ax-prover-base"
Repository = "https://github.com/Axiomatic-AI/ax-prover-base"
Issues = "https://github.com/Axiomatic-AI/ax-prover-base/issues"

[project.scripts]
ax-prover = "ax_prover.main:main"

[project.optional-dependencies]
dev = [
    "build>=1.0.0",
    "pre-commit>=4.5.0",
    "ruff>=0.14.6",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.4.0",
    "setuptools-scm>=8.0",
    "twine>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"ax_prover.configs" = ["*.yaml", ".env.secrets.example"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.setuptools_scm]
write_to = "src/ax_prover/_version.py"
fallback_version = "0.0.0.dev0"
